+++ app/Services/ProductService.php
@@ -1,18 +1,17 @@
 <?php
 namespace App\Services;
 
-use Cloudinary\Api\Exception\ApiError;
-use Illuminate\Support\Facades\Log;
-use Illuminate\Database\Eloquent\Collection;
-use Illuminate\Support\Facades\Auth;
-use Cloudinary\Cloudinary;
-use Illuminate\Support\Facades\Cache;
-use Illuminate\Support\Carbon;
-
+use App\Events\Product\ProductCreated;
 use App\Models\Filter;
 use App\Models\Product;
 use App\Models\Review;
 use App\Models\User;
+use Cloudinary\Api\Exception\ApiError;
+use Cloudinary\Cloudinary;
+use Illuminate\Database\Eloquent\Collection;
+use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Cache;
+use Illuminate\Support\Facades\Log;
 
 class ProductService {
     public function upload(array $validated_data, User $user) : void {
@@ -59,18 +58,7 @@ class ProductService {
             }
         }
 
-        activity('product')
-            ->causedBy($user)
-            ->performedOn($current_product)
-            ->withProperties([
-                'username' => $user->username,
-                'email' => $user->email,
-                'product_name' => $current_product->name,
-                'product_description' => $current_product->description,
-                'product_price' => $current_product->price,
-                'product_img_url' => $current_product->image_url,
-            ])
-        ->log('The user has successfully listed new product.');
+        event(new ProductCreated($current_product, $user));
     }
 
     public function sameProducts(int $id) : Collection {