+++ app/Services/UserService.php
@@ -1,11 +1,15 @@
 <?php
 namespace App\Services;
 
-use Illuminate\Support\Facades\Hash;
-
-use App\Models\User;
+use App\Events\User\UserChangedEmail;
+use App\Events\User\UserChangedUsername;
+use App\Events\User\UserError;
+use App\Events\User\UserSentReport;
+use App\Events\User\UserSentReview;
 use App\Models\Report;
 use App\Models\Review;
+use App\Models\User;
+use Illuminate\Support\Facades\Hash;
 
 
 class UserService {
@@ -19,21 +23,14 @@ class UserService {
             'confirmed' => false,
         ]);
 
-        activity('user')
-            ->causedBy($user)
-            ->withProperties([
-                'old_email' => $old_email,
-                'new_email' => $new_email,
-            ])
-        ->log('User has changed his email.');
+        event(new UserChangedEmail($user, $old_email, $new_email));
     }
 
     public function changeUsername (array $validated_data, User $user) : bool
     {
         if ($user->updated_at->diffInMinutes(now()) < 1) {
-            activity('user-error')
-                ->causedBy($user)
-            ->log('User change username failed: less than a minute has passed since the last update.');
+            $content = 'User change username failed: less than a minute has passed since the last update.';
+            event(new UserError($user, $content));
             return false;
         }
 
@@ -44,13 +41,7 @@ class UserService {
             'username' => $new_username,
         ]);
 
-        activity('user')
-            ->causedBy($user)
-            ->withProperties([
-                'old_username' => $old_username,
-                'new_username' => $new_username,
-            ])
-        ->log('User has changed his username.');
+        event(new UserChangedUsername($user, $old_username, $new_username));
         return true;
     }
 
@@ -60,9 +51,7 @@ class UserService {
             'password' => Hash::make($validated_data['password']),
         ]);
 
-        activity('user')
-            ->causedBy($user)
-        ->log('User has changed his password.');
+
     }
 
     public function sendReport(array $validatedData, User $user) : void
@@ -72,13 +61,7 @@ class UserService {
             'content' => $validatedData['content'],
         ]);
 
-        activity('user')
-            ->causedBy($user)
-            ->withProperties([
-                'username' => $user->username,
-                'content' => $validatedData['content'],
-            ])
-        ->log('Sent the report.');
+        event(new UserSentReport($user, $validatedData['content']));
     }
 
     public function sendReview(array $validatedData, User $user, int $id) : void
@@ -90,12 +73,6 @@ class UserService {
             'rating' => $validatedData['rating'],
         ]);
 
-        activity('user')
-            ->causedBy($user)
-            ->withProperties([
-                'username' => $user->username,
-                'message' => $validatedData['message'],
-            ])
-        ->log('Sent the review.');
+        event(new UserSentReview($user, $validatedData['message']));
     }
 }